Rename plugin methods
Renamed Methods:
addComponentExample -> addPluginComponentExample
addScreenExamples -> addPluginScreenExamples
removeComponentExample -> removePluginComponentExample
removeScreenExamples -> removePluginScreenExamples

Left old methods in place that call the new methods and print
deprecation warnings
