+++ packages/ignite-cli/src/extensions/ignite.js
@@ -7,11 +7,15 @@ const igniteConfigExt = require('./ignite/igniteConfig')
 const findIgnitePluginsExt = require('./ignite/findIgnitePlugins')
 const addModuleExt = require('./ignite/addModule')
 const removeModuleExt = require('./ignite/removeModule')
-const addScreenExamplesExt = require('./ignite/addScreenExamples')
-const removeScreenExamplesExt = require('./ignite/removeScreenExamples')
+const addScreenExamplesExt = require('./ignite/addScreenExamples')//Deprecated 3/2/17, Ignite 2 Beta
+const addPluginScreenExamplesExt = require('./ignite/addPluginScreenExamples')
+const removeScreenExamplesExt = require('./ignite/removeScreenExamples') //Deprecated 3/2/17, Ignite 2 Beta
+const removePluginScreenExamplesExt = require('./ignite/removePluginScreenExamples')
 const copyBatchExt = require('./ignite/copyBatch')
-const addComponentExampleExt = require('./ignite/addComponentExample')
-const removeComponentExampleExt = require('./ignite/removeComponentExample')
+const addComponentExampleExt = require('./ignite/addComponentExample') //Deprecated 3/2/17, Ignite 2 Beta
+const addPluginComponentExampleExt = require('./ignite/addPluginComponentExample')
+const removeComponentExampleExt = require('./ignite/removeComponentExample') //Deprecated 3/2/17, Ignite 2 Beta
+const removePluginComponentExampleExt = require('./ignite/removePluginComponentExample')
 const setDebugConfigExt = require('./ignite/setDebugConfig')
 const removeDebugConfigExt = require('./ignite/removeDebugConfig')
 const patchInFileExt = require('./ignite/patchInFile')
@@ -61,10 +65,14 @@ function attach (plugin, command, context) {
     addModule: addModuleExt(plugin, command, context),
     removeModule: removeModuleExt(plugin, command, context),
     copyBatch: copyBatchExt(plugin, command, context),
-    addComponentExample: addComponentExampleExt(plugin, command, context),
-    removeComponentExample: removeComponentExampleExt(plugin, command, context),
-    addScreenExamples: addScreenExamplesExt(plugin, command, context),
-    removeScreenExamples: removeScreenExamplesExt(plugin, command, context),
+    addComponentExample: addComponentExampleExt(plugin, command, context), //Deprecated 3/2/17, Ignite 2 Beta
+    addPluginComponentExample: addPluginComponentExampleExt(plugin, command, context),
+    removeComponentExample: removeComponentExampleExt(plugin, command, context), //Deprecated 3/2/17, Ignite 2 Beta
+    removePluginComponentExample: removePluginComponentExampleExt(plugin, command, context),
+    addScreenExamples: addScreenExamplesExt(plugin, command, context), //Deprecated 3/2/17, Ignite 2 Beta
+    addPluginScreenExamples: addPluginScreenExamplesExt(plugin, command, context),
+    removeScreenExamples: removeScreenExamplesExt(plugin, command, context), //Deprecated 3/2/17, Ignite 2 Beta
+    removePluginScreenExamples: removePluginScreenExamplesExt(plugin, command, context),
     setDebugConfig: setDebugConfigExt(plugin, command, context),
     removeDebugConfig: removeDebugConfigExt(plugin, command, context),
     patchInFile: patchInFileExt(plugin, command, context),
