+++ packages/ignite-cli/src/extensions/ignite/removeComponentExample.js
@@ -2,22 +2,11 @@ module.exports = (plugin, command, context) => {
   /**
    * Removes the component example.
    */
+  // DEPRECATED as of 3/2/17 as part of Ignite 2 Beta (https://github.com/infinitered/ignite/issues/636)
   function removeComponentExample (fileName) {
-    const { filesystem, patching, print } = context
-    print.info(`    ${print.checkmark} removing component example`)
-    // remove file from Components/Examples folder
-    filesystem.remove(
-      `${process.cwd()}/ignite/Examples/Components/${fileName}`
-    )
-    // remove reference in usage example screen (if it exists)
-    const destinationPath = `${process.cwd()}/ignite/DevScreens/PluginExamplesScreen.js`
-    if (filesystem.exists(destinationPath)) {
-      patching.replaceInFile(
-        destinationPath,
-        `import '../Examples/Components/${fileName}`,
-        ''
-      )
-    }
+    const { filesystem, patching, print, ignite } = context
+    print.warning('DEPRECATION_WARNING: Heads up! `ignite.removeComponentExample` is deprecated. Please use `ignite.removePluginComponentExample` instead.')
+    ignite.removePluginComponentExample(fileName)
   }
 
   return removeComponentExample
