+++ packages/ignite-cli/src/templates/plugin/index.js.ejs
@@ -12,7 +12,7 @@ const add = async function (context) {
   // install a npm module and link it
   await ignite.addModule(NPM_MODULE_NAME, { link: true })
 
-  <% if (props.answers.template === 'Yes') { %>await ignite.addComponentExample(EXAMPLE_FILE, { title: '<%= props.name %> Example' })<% } %>
+  <% if (props.answers.template === 'Yes') { %>await ignite.addPluginComponentExample(EXAMPLE_FILE, { title: '<%= props.name %> Example' })<% } %>
 
   // Example of copying templates/<%= props.name %> to App/<%= props.name %>
   // if (!filesystem.exists(`${APP_PATH}/App/<%= props.name %>`)) {
@@ -35,7 +35,7 @@ const remove = async function (context) {
   // remove the npm module and unlink it
   await ignite.removeModule(NPM_MODULE_NAME, { unlink: true })
 
-  <% if (props.answers.template === 'Yes') { %>await ignite.removeComponentExample(EXAMPLE_FILE)<% } %>
+  <% if (props.answers.template === 'Yes') { %>await ignite.removePluginComponentExample(EXAMPLE_FILE)<% } %>
 
   // Example of removing App/<%= props.name %> folder
   // const remove<%= props.name %> = await context.prompt.confirm(
